quarto

+++ lista_01/q4.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
-int media(int a, int b, int c, int d){
-    int media = (a+b+c+d)/4;
+float media(float a, float b, float c, float d){
+    float media = (a+b+c+d)/4;
     return media;
 }
 int main(){
-    int a,b,c,d;
-    scanf("%d %d %d %d",&a,&b,&c,&d);
-    printf("%d\n",media(a,b,c,d));
+    float a,b,c,d;
+    scanf("%f %f %f %f",&a,&b,&c,&d);
+    printf("%f\n",media(a,b,c,d));
     return 0;
 }
